Destroy the datasource when needed.
In PROD mode, the datasource should be destroyed on application stop.
In DEV mode, it should be destroyed before a new one is created.

Also, for JNDI datasources, a new property "db.destroyMethod" is added
to let decide if the application should destroy the datasource or not,
and how (By default, the datasource is not destroyed).