[#622] Moved the call to getString() outside of the try-catch block.
The try-catch block simply rewraps the exception with the message "Cannot parse
JSON", but if getString() throws an exception, then the problem had nothing to
do with parsing JSON. The new exception thrown by the try-catch block is a
RuntimeException, and that's pretty much all that getString() can throw, so it
seems fine to let that exception percolate up to the caller.

In practice, getString() may throw a java.net.ConnectException (connection
refused) exception.