[#541] UrlEncodedParser now obey hidden form-field '_charset_' when guessing which encoding is used in form POST
'_charset_' was invented by IE and is also handled by Firefox: when present in FORM,
these browsers update it with the actual encoding used when posting. Other browser don't handle it.

But as long as the form is rendered with the correct encoding in _charset_, Play! can read it and
use it when decoding form-posts using other encoding than the server default.

Important: Form must also use accept-charset with the same value to instruct the browser which encoding to use