[Multikeys]: Fixed invalid binding.test.html
Turns out that with refactoring a validation error that wasn't previously
reported is not reported:

In JPAPlugin.bind we try to find an entity from the DB if it has an ID, and
previously we didn't add any validation error if that entity was not found.
On the other hand, code that did the same thing but for relations like collections
or @*ToOne relations, DID produce an error on invalid IDs in GenericModel.
So toplevel invalid ID was no error, but nested invalid ID was error.
With the refactoring now both toplevel and nested entity binding from the DB
use the same code, which adds the error, making it more consistent.

Therefore, I have dropped the invalid ID from the binding tests to make
sure we did not get any error about that since we wanted to validate the
Color enum binding.

And I added a new test to make sure that invalid IDs do get us an error.