[#907] Changed avoidCascadeSaveLoops thread local to use a Set rather then a List.
Using a List for detecting save loops has severe performance implications -
approx O(n^2). Given a structure that had 2000 items in a list representing
a many-to-one relationship a save was taking approximately 90 seconds.
Switching to the Set dropped this to less that a second.

[#907] Test case for JPA cascade perfomance improments

New test model and case has been added to verify the CascadeType.ALL cases for
insert, update and delete.