[#1395] Fix orphan removal bug caused by bad binding of collection properties
Now when a collection property is going to binded instead of overwrite it's value with a new collection. The original collection is obtained and cleared. So if the collection is a persistent one the value are merged correctly. It also performs a merge of the object in the JPA Plugin so now it is not necesary to merge the object in the controller.