[#1466] Remove cookies with maxAge from savedCookies
Cookies with maxAge 0 should be deleted by the browser. Scope.Flash.save() uses this to
clear the PLAY_FLASH cookie.

FunctionalTest simply ignores cookies with maxAge 0 completely. This patch ensures that
a cookie with maxAge 0 will delete a previous cookie if it exists.

The cookie emulation still ignores paths, so theoretically a maxAge zero cookie for path
/foo would delete a cookie for path /, but in practice this should not occur in test cases.