Fix redirect to wrong domain and port behind apache
Behind apache reverce proxy that has ProxyPreserveHost off,
when XForwardedSupport is specified in application.conf,
the request domain and port are not changed to original domain and port.
As a result, redirect to wrong url which is created by the request domain and port.

To fix the problem, XForwardedOverwriteDomainAndPort option is added.
When "XForwardedOverwriteDomainAndPort=on" is set in application.conf,
the request domain and port are changed to original domain and port.