Updata Controller javadoc to inform that stream content may be fully loaded in memory
Added the following comment on all renderBinary methods that accept an InputStream but no length parameter:
"Content is fully loaded in memory, so it should not be used with large data."

See the RenderBinary class: all data is copied into a ByteArrayInputStream (response.out). This can lead to OutOfMemoryErrors, so I think should be documented.