Clarify enhancement verification
- Renamed isFrameworkClass() to isFrameworkCode(). The check for framework code
includes JDK classes. We see these in the stack as the Play framework is reflectively
invoking application code.

- Pulled the guard clause that detects that we are no longer examing application code
away from the enhancement check to highlight that it is not merely checking if the
current class should be enhanced but instead signalling the entire verification is
complete and that we should exit the method.