Fix issue #1463 by restoring behavior --timeout parameter
The upgrade to python3 broke the processing of the --timeout parameter
due to a type mismatch (a string was compared to an integer).  The fix
is to always treat the --timeout parameter as a string, deferring to the
test runner to parse it and report errors.

This also fixes a typo in the name of a variable "webclient_timeout"
(it was named "weblcient_timeout").

It also adds some new regression tests for the auto-test functionality.